requirements.txt
requests
openai
python-dotenv
flask
langchain
langchain-openai
langsmith